examples-python/gemini-stream-sse-connector/requirements.txt
# Vastar Connector SDK (local path)
-e ../../sdk-python

# Configuration
pyyaml>=6.0
numpy
